pyproject.toml
[tool.poetry]
name = "flaskapi"
version = "0.1.0"
description = ""
authors = ["adrianfhr <adrianfhrr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flask-restful = "^0.3.10"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
psycopg2 = "^2.9.9"
flask-migrate = "^4.0.7"
flask-marshmallow = "^1.2.1"
marshmallow-sqlalchemy = "^1.1.0"
passlib = "^1.7.4"
flask-jwt-extended = "^4.6.0"


[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
